Basic MQTT implementation.
This is completely untested. It compiles, but one of my rules is not
to check that my code works when I want to go to bed, or I'll be up
all night. Get the code to a state where it is ready to be tested,
then call it a night.