Change flasher config to not auto reset.
For some reason, after a month of being away from this, my dev board
no longer wants to automatically flash. Not sure what's going on
there. But, the workaround appears to be to tell it to flash then hold
the flash button, press and release the reset button, then release the
flash button, and then it's happy. But, this only works if it doesn't
try to reset the chip when it first connects.