Fix a bug where a too short sleep would fail to connect.
It turns out that the guard is unecessary - we can both call connect
when connected and, more importantly, we need to connect when we're in
the middle of stopping the stack if we want to restart the WiFi.

This was causing an issue in the case of doing tests where I was just
sending temperature data without a long delay - the WiFi wouldn't
finish coming down so it would fail to send every other packet.

With this guard remove, we connect as normal and send data just fine.