Fix a bug with our sleep interval rolling over.
If we took too long to send data and therefore our next target waking
time was before now, we'd subtract now from it, which means it would
go negative and then be used to calculate the interval which would
roll over but, since it's unsigned, we'd end up with a delay close to
2^32 microseconds, which is still a really long time.

This revised code now checks to see if we missed our wake up time and,
if so, just runs our data acquisition loop again.

Note that if this happens often, it means that the system is going to
drain the battery very quickly because it never goes to sleep - it's
just constantly awake waiting WiFi / data send / etc.

Now, that said, it shouldn't happen very often - my tests show that it
takes about 7.5 seconds to connect to WiFi, get a DHCP lease, do the
DNS lookup for the COAP server, get the temperature, send it,
disconnect, tear everything down and go to sleep. So, unless your
desired periodicity is something like 5 seconds, it will be fine,
given that the design intent is to have the update rate commonly be no
faster than once per minute.